wip(refactor): fix test in gameWithScore
