--- src/main/scala/scatan/model/game/state/FinalState.scala
@@ -1,5 +1,0 @@
-package scatan.model.game.state
-
-import scatan.lib.game.Player
-
-trait FinalState extends ScoreKnowledge[FinalState]
